Raycasting improvements by mortinso (#6)
Simplified some stuff,
Draws corners
Draws doorways properly and it responds correctly to player movement
No fish eyed camera
Smooth lines
