+++ inc/cub3D.h
@@ -6,15 +6,15 @@
 /*   By: mortins- <mortins-@student.42lisboa.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/14 15:31:01 by mortins-          #+#    #+#             */
-/*   Updated: 2024/08/22 17:57:52 by mortins-         ###   ########.fr       */
+/*   Updated: 2024/09/02 20:21:14 by mortins-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef CUB3D_H
 # define CUB3D_H
 
-# define SCREEN_W 2560
-# define SCREEN_H 1440
+# define SCREEN_W 1920
+# define SCREEN_H 1080
 
 //----------------------------------INCLUDES------------------------------------
 # include <stdio.h>
@@ -27,10 +27,10 @@
 
 # define PI 3.1415926535
 # define RAD_DEGREE 0.0174533
-# define FOV 90
-# define CELL 64 //2D
+# define CELL 64
 
 // v-------- 2Dview -------v
+# define FOV 90
 # define P_SIZE 15 //2D
 # define P_COLOR 0x00ff0000
 # define A_COLOR 0x00ffff00
@@ -76,19 +76,21 @@ typedef struct s_raycast
 {
 	t_vector	line_dist;
 	t_vector	delta_dist;
+	t_vector	ray_dir;
 	int			step_x;
 	int			step_y;
 	int			map_x;
 	int			map_y;
 	int			orientation;
 	double		wall_dist;
-	int			screen_x;
+	double		win_x;
 }	t_raycast;
 
 typedef struct s_player
 {
 	t_vector	pos;
 	t_vector	dir;
+	t_vector	plane;
 }	t_player;
 
 typedef struct s_map
@@ -129,9 +131,6 @@ void	init(t_cube *cube, char *map_fd);
 // main.c
 void	my_mlx_pixel_put(t_img *data, int x, int y, int color);
 
-// raycasting.c
-void	raycasting(t_cube *cube, t_vector dir);
-
 // utils.c
 size_t	array_size(char **array);
 void	free_array(char **array);
@@ -156,5 +155,13 @@ void	map_error(t_cube *cube, int fd, char *error_msg);
 int		is_texture_identifier(const char *line);
 void	purge_textures(t_cube *cube);
 void	set_texture(t_cube *cube, int fd, char *line);
+
+// +++++++++++++++ raycasting/[.....] +++++++++++++++++++++++++++++++++++++++++
+// drawing.c
+void	draw_wall(t_cube *cube, t_raycast *cast, int x);
+void	draw_floor(t_cube *cube);
+
+// raycasting.c
+void	raycasting(t_cube *cube, t_raycast *cast, t_vector dir);
 
 #endif
