+++ src/main.c
@@ -6,7 +6,7 @@
 /*   By: mortins- <mortins-@student.42lisboa.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/12 16:27:16 by mortins-          #+#    #+#             */
-/*   Updated: 2024/08/20 18:22:51 by mortins-         ###   ########.fr       */
+/*   Updated: 2024/09/02 20:22:25 by mortins-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,9 +24,8 @@ void	my_mlx_pixel_put(t_img *data, int x, int y, int color)
 // Function that draws to the window
 int	render_frame(t_cube *cube)
 {
-	mlx_put_image_to_window(cube->mlx, cube->window_2, cube->minimap.img, 0, 0);
-	draw_square(cube, P_SIZE, cube->player.pos, P_COLOR);
-	raycasting(cube, cube->player.dir);
+	/* mlx_put_image_to_window(cube->mlx, cube->window_2, cube->minimap.img, 0, 0);
+	draw_square(cube, P_SIZE, cube->player.pos, P_COLOR); */
 	mlx_put_image_to_window(cube->mlx, cube->game_window, cube->game.img, 0, 0);
 	//draw_fov(cube);
 	return (0);
@@ -55,7 +54,8 @@ int	main(int argc, char **argv)
 		ft_printf("Window failed\n");
 		exit (1);
 	}
-	cube.window_2 = mlx_new_window(cube.mlx, SCREEN_W, SCREEN_H, "minimap");
+	raycasting(&cube, &cube.raycast, cube.player.dir);
+	/* cube.window_2 = mlx_new_window(cube.mlx, SCREEN_W, SCREEN_H, "minimap");
 	if (!cube.window_2)
 	{
 		mlx_clear_window(cube.mlx, cube.game_window);
@@ -64,10 +64,10 @@ int	main(int argc, char **argv)
 		free(cube.mlx);
 		ft_printf("Window failed\n");
 		exit (1);
-	}
+	} */
 	mlx_hook(cube.game_window, 2, 1L << 0, keypress, &cube);
 	mlx_hook(cube.game_window, 17, 0L, destruct, &cube);
-	mlx_hook(cube.window_2, 17, 0L, destruct, &cube);
+	// mlx_hook(cube.window_2, 17, 0L, destruct, &cube);
 	mlx_loop_hook(cube.mlx, render_frame, &cube);
 	mlx_loop(cube.mlx);
 }
